+++ src/src/beneficiaries/beneficiaries.service.ts
@@ -15,6 +15,7 @@ import { UserHelperService } from '../helper/userHelper.service';
 import { HasuraService as HasuraServiceFromServices } from '../services/hasura/hasura.service';
 import { KeycloakService } from '../services/keycloak/keycloak.service';
 import { UploadFileService } from 'src/upload-file/upload-file.service';
+import { BeneficiariesCoreService } from './beneficiaries.core.service';
 @Injectable()
 export class BeneficiariesService {
 	public url = process.env.HASURA_BASE_URL;
@@ -30,6 +31,7 @@ export class BeneficiariesService {
 		private configService: ConfigService,
 		private enumService: EnumService,
 		private uploadFileService: UploadFileService,
+		private beneficiariesCoreService: BeneficiariesCoreService,
 	) {}
 
 	allStatus = this.enumService.getEnumValue('BENEFICIARY_STATUS').data;
@@ -1599,66 +1601,77 @@ export class BeneficiariesService {
 	}
 
 	public async statusUpdate(body: any, request: any) {
-		const { data: updatedUser } = await this.userById(body?.user_id);
-		const allStatuses = this.enumService
-			.getEnumValue('BENEFICIARY_STATUS')
-			.data.map((enumData) => enumData.value);
+		const status_response =
+			await this.beneficiariesCoreService.statusUpdate(
+				body,
 
-		if (!allStatuses.includes(body.status)) {
-			return {
-				status: 400,
-				success: false,
-				message: `Invalid status`,
-				data: {},
-			};
-		}
-		const res = await this.hasuraService.update(
-			updatedUser?.program_beneficiaries?.id,
-			'program_beneficiaries',
-			{
-				...body,
-				reason_for_status_update: body.reason_for_status_update?.trim()
-					? body.reason_for_status_update?.trim()
-					: body.status,
-			},
-			this.returnFields,
-			[...this.returnFields, 'id'],
-		);
+				request,
+			);
 
-		const newdata = (
-			await this.userById(res?.program_beneficiaries?.user_id)
-		).data;
-
-		const audit = await this.userService.addAuditLog(
-			body?.user_id,
-			request.mw_userid,
-			'program_beneficiaries.status',
-			updatedUser?.program_beneficiaries?.id,
-			{
-				status: updatedUser?.program_beneficiaries?.status,
-				reason_for_status_update:
-					updatedUser?.program_beneficiaries
-						?.reason_for_status_update,
-			},
-			{
-				status: newdata?.program_beneficiaries?.status,
-				reason_for_status_update:
-					newdata?.program_beneficiaries?.reason_for_status_update,
-			},
-			['status', 'reason_for_status_update'],
-		);
+		console.log('status_response', status_response);
 
 		return {
 			status: 200,
 			success: true,
 			message: 'Status Updated successfully!',
-			data: (await this.userById(res?.program_beneficiaries?.user_id))
-				.data,
+			data: (
+				await this.beneficiariesCoreService.userById(
+					status_response?.program_beneficiaries?.user_id,
+				)
+			).data,
+		};
+	}
+
+	public async statusUpdateByIp(body: any, request: any) {
+		const user = await this.userService.ipUserInfo(request);
+		if (!user?.data?.program_users?.[0]?.organisation_id) {
+			return {
+				success: false,
+				message: 'Invalid Ip',
+				data: {},
+			};
+		}
+		let organisation_id = user?.data?.program_users?.[0]?.organisation_id;
+
+		//check validation for id benlongs to same IP under prerak
+		let data = {
+			query: `query MyQuery {
+				users(where: {program_faciltators: {parent_ip: {_eq: "${organisation_id}"}, beneficiaries: {user_id: {_eq: ${body?.user_id}}}}}){
+				  program_faciltators{
+					parent_ip
+				  }
+				}
+			  }`,
+		};
+
+		const result = await this.hasuraServiceFromServices.getData(data);
+
+		if (result?.data?.users?.length == 0) {
+			return {
+				status: 401,
+				success: false,
+				message: 'IP_ACCESS_DENIED',
+				data: {},
+			};
+		}
+		const status_response =
+			await this.beneficiariesCoreService.statusUpdate(body, request);
+
+		return {
+			status: 200,
+			success: true,
+			message: 'Status Updated successfully!',
+			data: (
+				await this.beneficiariesCoreService.userById(
+					status_response?.program_beneficiaries?.user_id,
+				)
+			).data,
 		};
 	}
 
 	public async setEnrollmentStatus(body: any, request: any) {
-		const { data: updatedUser } = await this.userById(body.user_id);
+		const { data: updatedUser } =
+			await this.beneficiariesCoreService.userById(body.user_id);
 		const allEnrollmentStatuses = this.enumService
 			.getEnumValue('ENROLLEMENT_VERIFICATION_STATUS')
 			.data.map((enumData) => enumData.value);
@@ -1739,7 +1752,9 @@ export class BeneficiariesService {
 		}
 
 		const newdata = (
-			await this.userById(res?.program_beneficiaries?.user_id)
+			await this.beneficiariesCoreService.userById(
+				res?.program_beneficiaries?.user_id,
+			)
 		).data;
 
 		const audit = await this.userService.addAuditLog(
@@ -1764,8 +1779,11 @@ export class BeneficiariesService {
 			status: 200,
 			success: true,
 			message: 'Status Updated successfully!',
-			data: (await this.userById(res?.program_beneficiaries?.user_id))
-				.data,
+			data: (
+				await this.beneficiariesCoreService.userById(
+					res?.program_beneficiaries?.user_id,
+				)
+			).data,
 		};
 	}
 
@@ -1823,7 +1841,8 @@ export class BeneficiariesService {
 
 	async create(req: any, request, response, update = false) {
 		const user = await this.userService.ipUserInfo(request);
-		const { data: beneficiaryUser } = await this.userById(req.id);
+		const { data: beneficiaryUser } =
+			await this.beneficiariesCoreService.userById(req.id);
 		if (!beneficiaryUser) {
 			return response.status(400).json({
 				success: false,
@@ -2780,7 +2799,8 @@ export class BeneficiariesService {
 					update,
 				);
 
-				const { data: updatedUser } = await this.userById(req.id);
+				const { data: updatedUser } =
+					await this.beneficiariesCoreService.userById(req.id);
 				if (updatedUser.program_beneficiaries.enrollment_number) {
 					let status = null;
 					let reason = null;
@@ -2866,7 +2886,8 @@ export class BeneficiariesService {
 				break;
 			}
 		}
-		const { data: updatedUser } = await this.userById(user_id);
+		const { data: updatedUser } =
+			await this.beneficiariesCoreService.userById(user_id);
 		return response.status(200).json({
 			success: true,
 			message: 'User data fetched successfully!',
@@ -2897,168 +2918,19 @@ export class BeneficiariesService {
 				['device_ownership', 'device_type', 'user_id'],
 			);
 		}
-		return await this.userById(user_id);
+		return await this.beneficiariesCoreService.userById(user_id);
 	}
 
-	async userById(id: any) {
-		const data = {
-			query: `query searchById {
-            users_by_pk(id: ${id}) {
-			aadhaar_verification_mode
-			aadhar_no
-			aadhar_token
-			aadhar_verified
-			address
-			address_line_1
-			address_line_2
-			alternative_mobile_number
-			block
-			block_id
-			block_village_id
-			created_at
-			created_by
-			district
-			district_id
-			dob
-			duplicate_reason
-			email_id
-			email_verified
-			first_name
-			gender
-			grampanchayat
-			id
-			is_duplicate
-			keycloak_id
-			last_name
-			lat
-			long
-			middle_name
-			mobile
-			mobile_no_verified
-			pincode
-			profile_photo_1
-			profile_photo_2
-			profile_photo_3
-			profile_url
-			state
-			state_id
-			updated_at
-			updated_by
-			village
-			username
-            program_beneficiaries {
-            beneficiaries_found_at
-            created_by
-            facilitator_id
-			original_facilitator_id
-            id
-            status
-            reason_for_status_update
-            academic_year_id
-            user_id
-            enrollment_number
-            enrollment_status
-            enrolled_for_board
-            type_of_enrollement
-            subjects
-            payment_receipt_document_id
-            program_id
-            updated_by
-            documents_status
-            learning_motivation
-            type_of_support_needed
-			learning_level
-			enrollment_date
-			enrollment_first_name
-			enrollment_middle_name
-			enrollment_last_name
-			enrollment_dob
-			enrollment_aadhaar_no
-			is_eligible
-			enrollment_verification_reason
-			enrollment_verification_status
-			document {
-				context
-				context_id
-				created_by
-				document_sub_type
-				doument_type
-				id
-				name
-				path
-				provider
-				updated_by
-				user_id
-			  }
-          }
-            core_beneficiaries {
-            career_aspiration
-            updated_by
-            type_of_learner
-            status
-			type_of_enrollement
-            reason_of_leaving_education
-            previous_school_type
-            mobile_ownership
-            learner_wish_to_pursue_education
-            last_standard_of_education_year
-            last_standard_of_education
-            last_school_type
-            id
-            connect_via_refrence
-            created_by
-            device_ownership
-            device_type
-            document_id
-            enrolled_for_board
-            enrollement_status
-            father_first_name
-            father_middle_name
-            father_last_name
-            mother_first_name
-            mother_middle_name
-            mother_last_name
-            career_aspiration_details
-            alternative_device_ownership
-            alternative_device_type
-            mark_as_whatsapp_number
-          }
-          references {
-            id
-            name
-            first_name
-            last_name
-            middle_name
-            relation
-            contact_number
-            designation
-            document_id
-            type_of_document
-            context
-            context_id
-          }
-          extended_users {
-            marital_status
-            designation
-            created_by
-            id
-            user_id
-            updated_by
-            social_category
-            qualification_id
-          }
-        }}`,
-		};
-		const response = await this.hasuraServiceFromServices.getData(data);
-		let result = response?.data?.users_by_pk;
-		if (result) {
-			result.program_beneficiaries = result?.program_beneficiaries?.[0];
-		}
-		return {
-			message: 'User data fetched successfully.',
-			data: result,
-		};
-	}
+	// async userById(id: any) {
+	// 	const result = await this.beneficiariesCoreService.userById(id);
+	// 	if (result) {
+	// 		result.program_beneficiaries = result?.program_beneficiaries?.[0];
+	// 	}
+	// 	return {
+	// 		message: 'User data fetched successfully.',
+	// 		data: result,
+	// 	};
+	// }
 
 	public async getAllDuplicatesUnderIp(
 		id: number,
@@ -3355,7 +3227,9 @@ export class BeneficiariesService {
 			}
 		}
 
-		const beneficiaryDetails = (await this.userById(beneficiaryId)).data;
+		const beneficiaryDetails = (
+			await this.beneficiariesCoreService.userById(beneficiaryId)
+		).data;
 
 		const updatePayload: any = {
 			facilitator_id: newFacilitatorId,
